Code/requirements.txt
keras==2.15.0
matplotlib==3.8.2
moviepy==1.0.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
pyserial==3.5
tensorflow==2.15.0
tensorflow_cpu_aws==2.15.0
tflite_runtime==2.14.0
